test: extract integration test helper
